// src/openapi/client/models/SimpleAutoBrand.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My App
 * This is an awesome app!
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SimpleAutoBrand
 */
export interface SimpleAutoBrand {
    /**
     * 
     * @type {number}
     * @memberof SimpleAutoBrand
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof SimpleAutoBrand
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof SimpleAutoBrand
     */
    orderNumber: number;
    /**
     * 
     * @type {string}
     * @memberof SimpleAutoBrand
     */
    image: string;
    /**
     * 
     * @type {string}
     * @memberof SimpleAutoBrand
     */
    imageFilePath: string;
}

/**
 * Check if a given object implements the SimpleAutoBrand interface.
 */
export function instanceOfSimpleAutoBrand(value: object): value is SimpleAutoBrand {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('orderNumber' in value) || value['orderNumber'] === undefined) return false;
    if (!('image' in value) || value['image'] === undefined) return false;
    if (!('imageFilePath' in value) || value['imageFilePath'] === undefined) return false;
    return true;
}

export function SimpleAutoBrandFromJSON(json: any): SimpleAutoBrand {
    return SimpleAutoBrandFromJSONTyped(json, false);
}

export function SimpleAutoBrandFromJSONTyped(json: any, ignoreDiscriminator: boolean): SimpleAutoBrand {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'orderNumber': json['orderNumber'],
        'image': json['image'],
        'imageFilePath': json['imageFilePath'],
    };
}

export function SimpleAutoBrandToJSON(json: any): SimpleAutoBrand {
    return SimpleAutoBrandToJSONTyped(json, false);
}

export function SimpleAutoBrandToJSONTyped(value?: SimpleAutoBrand | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'orderNumber': value['orderNumber'],
        'image': value['image'],
        'imageFilePath': value['imageFilePath'],
    };
}