// src/openapi/client/models/SimpleAutoModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My App
 * This is an awesome app!
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SimpleAutoModel
 */
export interface SimpleAutoModel {
    /**
     * 
     * @type {number}
     * @memberof SimpleAutoModel
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof SimpleAutoModel
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof SimpleAutoModel
     */
    orderNumber: number;
}

/**
 * Check if a given object implements the SimpleAutoModel interface.
 */
export function instanceOfSimpleAutoModel(value: object): value is SimpleAutoModel {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('orderNumber' in value) || value['orderNumber'] === undefined) return false;
    return true;
}

export function SimpleAutoModelFromJSON(json: any): SimpleAutoModel {
    return SimpleAutoModelFromJSONTyped(json, false);
}

export function SimpleAutoModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): SimpleAutoModel {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'orderNumber': json['orderNumber'],
    };
}

export function SimpleAutoModelToJSON(json: any): SimpleAutoModel {
    return SimpleAutoModelToJSONTyped(json, false);
}

export function SimpleAutoModelToJSONTyped(value?: SimpleAutoModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'orderNumber': value['orderNumber'],
    };
}